Moved session-key constants
