--- gui/src/constants/session-key.ts
@@ -1,7 +1,0 @@
-export enum SessionKey {
-  intermediateStub = "intermediateStub",
-  settings = "settings",
-  userToken = "userToken",
-  stubsFilter = "stubsFilter",
-  requestsFilter = "requestsFilter",
-}
